libs/shared/backend/users/data-access/src/lib/users.repository.ts
import { Injectable, NotFoundException } from '@nestjs/common';
import { InjectRepository } from '@nestjs/typeorm';
import { FindManyOptions, Repository } from 'typeorm';
import { User } from './user.entity';

@Injectable()
export class UsersRepository {
  constructor(
    @InjectRepository(User) private readonly userRepository: Repository<User>,
  ) {}

  findAll(options?: FindManyOptions<User>): Promise<User[]> {
    return this.userRepository.find(options);
  }

  async findOne(id: string): Promise<User> {
    const user = await this.userRepository.findOne({ where: { id } });
    if (!user) {
      throw new NotFoundException(`User #${id} not found`);
    }
    return user;
  }

  async create(user: Partial<User>): Promise<User> {
    const newUser = this.userRepository.create(user);
    return this.userRepository.save(newUser);
  }

  async update(id: string, user: Partial<User>): Promise<User> {
    const existingUser = await this.userRepository.preload({
      id,
      ...user,
    });
    if (!existingUser) {
      throw new NotFoundException(`User #${id} not found`);
    }
    return this.userRepository.save(existingUser);
  }

  async delete(id: string): Promise<string> {
    const user = await this.findOne(id);
    await this.userRepository.remove(user);
    return id;
  }
}
